j'ai modifier le diagramme de class et aussi j'ai ajoute les fonctionnalites de carte bank

+++ E_Bank_Solution/src/main/java/com/example/E_Bank_Solution/Repository/CartebankRepository.java
@@ -3,5 +3,5 @@ package com.example.E_Bank_Solution.Repository;
 import com.example.E_Bank_Solution.Model.Cartebank;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CartebankRepository extends JpaRepository <Cartebank, Integer> {
+public interface CartebankRepository extends JpaRepository <Cartebank, Long> {
 }
